@@ -8,8 +8,8 @@
     "@fortawesome/free-solid-svg-icons": "^5.9.0",
     "@fortawesome/react-fontawesome": "^0.1.4",
     "node-sass": "^4.11.0",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.0.5",
     "react-scroll": "^1.7.13",
