@@ -1,9 +1,9 @@
 import React from "react";
 import styled from "styled-components";
 
-const Wrapper = ({ className, title, children, midHeight, id, last }) => {
+const Wrapper = ({ className, title, children, midHeight, id}) => {
     return (
-        <div className={className} midHeight={midHeight} last={last}>
+        <div className={className} midHeight={midHeight}>
             <div style={{ paddingTop: '44px' }} id={id}></div>
             <h1 style={{ textAlign:'center', fontWeight:'bold', fontSize:'1.5rem', marginBottom:'5%'}}>{title}</h1>
             {children}
